// backend/src/api/auth/constants/validationMessages.ts
export enum ValidationMessages {
  NAME_REQUIRED = "Name is required",
  NAME_MAX_LENGTH = "Name must not exceed 50 characters",
  INVALID_EMAIL = "Invalid email format",
  PASSWORD_REQUIRED = "Password is required",
  PASSWORD_MIN_LENGTH = "Password must be at least 6 characters long",
  PASSWORD_STRONG = "Password must include at least one uppercase letter, one number, and one lowercase letter",
}